fin primera parte

+++ app/Http/Controllers/Instructor/CourseController.php
@@ -121,7 +121,10 @@ class CourseController extends Controller
         $course->status = 2;
         $course->save();
 
-        $course->observation->delete();
+        if ($course->observation) {
+            $course->observation->delete();
+        }
+        
         return redirect()->route('instructor.courses.edit',$course);
     }
